+++ js/MediaStream.js
@@ -32,7 +32,8 @@ function newMediaStreamId() {
 }
 
 // Save original MediaStream
-var originalMediaStream = window.MediaStream;
+var originalMediaStream = window.MediaStream || window.Blob;
+//var originalMediaStream = window.Blob;
 var originalMediaStreamTrack = MediaStreamTrack.originalMediaStreamTrack;
 
 /**
@@ -45,8 +46,9 @@ function MediaStream(arg, id) {
 	// new MediaStream(originalMediaStream) // stream
 	// new MediaStream(originalMediaStreamTrack[]) // tracks
 	if (
-		(arg instanceof originalMediaStream && typeof arg.getBlobId === 'undefined') ||
-			(Array.isArray(arg) && arg[0] instanceof originalMediaStreamTrack)
+		!(arg instanceof window.Blob) &&
+			(arg instanceof originalMediaStream && typeof arg.getBlobId === 'undefined') ||
+				(Array.isArray(arg) && arg[0] instanceof originalMediaStreamTrack)
 	) {
 		return new originalMediaStream(arg);
 	}
@@ -55,9 +57,26 @@ function MediaStream(arg, id) {
 	// new MediaStream(MediaStreamTrack[]) // tracks
 	// new MediaStream() // empty
 
-	// TODO attempt CustomMediaStream extend.
+	id = id || newMediaStreamId();
+	var blobId = 'MediaStream_' + id;
+
 	// Extend returned MediaTream with custom MediaStream
-	var stream = new (Function.prototype.bind.apply(originalMediaStream.bind(this), [])); // jshint ignore:line
+	var stream;
+	if (originalMediaStream !== window.Blob) {
+		stream = new (Function.prototype.bind.apply(originalMediaStream.bind(this), [])); // jshint ignore:line
+
+	// Fallback on Blob if originalMediaStream is not a MediaStream and Emulate EventTarget
+	} else {
+		stream = new Blob([blobId], {
+			type: 'stream'
+		});
+
+		var target = document.createTextNode(null);
+		stream.addEventListener = target.addEventListener.bind(target);
+		stream.removeEventListener = target.removeEventListener.bind(target);
+		stream.dispatchEvent = target.dispatchEvent.bind(target);
+	}
+
 	Object.defineProperties(stream, Object.getOwnPropertyDescriptors(MediaStream.prototype));
 
 	// Make it an EventTarget.
@@ -78,7 +97,7 @@ function MediaStream(arg, id) {
 	stream._videoTracks = {};
 
 	// Store the stream into the dictionary.
-	stream._blobId = 'MediaStream_' + stream.id;
+	stream._blobId = blobId;
 	mediaStreams[stream._blobId] = stream;
 
 	// Convert arg to array of tracks if possible
@@ -243,7 +262,6 @@ MediaStream.prototype.getTrackById = function (id) {
 	return this._audioTracks[id] || this._videoTracks[id] || null;
 };
 
-
 MediaStream.prototype.addTrack = function (track) {
 	debug('addTrack() [track:%o]', track);
 
@@ -266,8 +284,11 @@ MediaStream.prototype.addTrack = function (track) {
 	addListenerForTrackEnded.call(this, track);
 
 	exec(null, null, 'iosrtcPlugin', 'MediaStream_addTrack', [this.id, track.id]);
-};
 
+	this.dispatchEvent(new Event('update'));
+
+	this.emitConnected();
+};
 
 MediaStream.prototype.removeTrack = function (track) {
 	debug('removeTrack() [track:%o]', track);
@@ -290,13 +311,20 @@ MediaStream.prototype.removeTrack = function (track) {
 
 	exec(null, null, 'iosrtcPlugin', 'MediaStream_removeTrack', [this.id, track.id]);
 
+	this.dispatchEvent(new Event('update'));
+
 	checkActive.call(this);
 };
 
 
 MediaStream.prototype.clone = function () {
-	debug('clone()');
-	return new MediaStream(this);
+
+	var newStream = MediaStream();
+	this.getTracks().forEach(function (track) {
+		newStream.addTrack(track.clone());
+	});
+
+	return newStream;
 };
 
 // Backwards compatible API.
@@ -458,6 +486,7 @@ function onEvent(data) {
 			event.track = track;
 
 			this.dispatchEvent(event);
+
 			// Also emit 'update' for the MediaStreamRenderer.
 			this.dispatchEvent(new Event('update'));