+++ src/PluginMediaStream.swift
@@ -8,33 +8,33 @@ class PluginMediaStream : NSObject {
 	var videoTracks: [String : PluginMediaStreamTrack] = [:]
 	var eventListener: ((_ data: NSDictionary) -> Void)?
 	var eventListenerForAddTrack: ((_ pluginMediaStreamTrack: PluginMediaStreamTrack) -> Void)?
-	var eventListenerForRemoveTrack: ((_ id: String) -> Void)?
+	var eventListenerForRemoveTrack: ((_ pluginMediaStreamTrack: PluginMediaStreamTrack) -> Void)?
 
 	/**
 	 * Constructor for pc.onaddstream event and getUserMedia().
 	 */
-	init(rtcMediaStream: RTCMediaStream) {
+	init(rtcMediaStream: RTCMediaStream, streamId: String? = nil) {
 		NSLog("PluginMediaStream#init()")
 
-		self.rtcMediaStream = rtcMediaStream
+		self.rtcMediaStream = rtcMediaStream;
 
-		// Handle possible duplicate remote streamId with janus name
-		// See: https://github.com/cordova-rtc/cordova-plugin-iosrtc/issues/432
-		if (rtcMediaStream.streamId.starts(with: "janus")) {
+		if (streamId == nil) {
+			// Handle possible duplicate remote trackId with  janus or short duplicate name
+			// See: https://github.com/cordova-rtc/cordova-plugin-iosrtc/issues/432
 			self.id = rtcMediaStream.streamId + "_" + UUID().uuidString;
 		} else {
-			self.id = rtcMediaStream.streamId;
+			self.id = streamId!;
 		}
 
 		for track: RTCMediaStreamTrack in (self.rtcMediaStream.audioTracks as Array<RTCMediaStreamTrack>) {
-			let pluginMediaStreamTrack = PluginMediaStreamTrack(rtcMediaStreamTrack: track, streamId: id)
+			let pluginMediaStreamTrack = PluginMediaStreamTrack(rtcMediaStreamTrack: track)
 
 			pluginMediaStreamTrack.run()
 			self.audioTracks[pluginMediaStreamTrack.id] = pluginMediaStreamTrack
 		}
 
 		for track: RTCMediaStreamTrack in (self.rtcMediaStream.videoTracks as Array<RTCMediaStreamTrack>) {
-			let pluginMediaStreamTrack = PluginMediaStreamTrack(rtcMediaStreamTrack: track, streamId: id)
+			let pluginMediaStreamTrack = PluginMediaStreamTrack(rtcMediaStreamTrack: track)
 
 			pluginMediaStreamTrack.run()
 			self.videoTracks[pluginMediaStreamTrack.id] = pluginMediaStreamTrack
@@ -43,20 +43,26 @@ class PluginMediaStream : NSObject {
 
 	deinit {
 		NSLog("PluginMediaStream#deinit()")
-		for (id, _) in audioTracks {
-			if(self.eventListenerForRemoveTrack != nil) {
-				self.eventListenerForRemoveTrack!(id)
-			}
-		}
-		for (id, _) in videoTracks {
-			if(self.eventListenerForRemoveTrack != nil) {
-				self.eventListenerForRemoveTrack!(id)
-			}
-		}
+		stop();
 	}
 
 	func run() {
 		NSLog("PluginMediaStream#run()")
+	}
+	
+	func stop() {
+		NSLog("PluginMediaStream#stop()")
+		
+		for (_, track) in audioTracks {
+			if(self.eventListenerForRemoveTrack != nil) {
+				self.eventListenerForRemoveTrack!(track)
+			}
+		}
+		for (_, track) in videoTracks {
+			if(self.eventListenerForRemoveTrack != nil) {
+				self.eventListenerForRemoveTrack!(track)
+			}
+		}
 	}
 
 	func getJSON() -> NSDictionary {
@@ -80,7 +86,7 @@ class PluginMediaStream : NSObject {
 	func setListener(
 		_ eventListener: @escaping (_ data: NSDictionary) -> Void,
 		eventListenerForAddTrack: ((_ pluginMediaStreamTrack: PluginMediaStreamTrack) -> Void)?,
-		eventListenerForRemoveTrack: ((_ id: String) -> Void)?
+		eventListenerForRemoveTrack: ((_ pluginMediaStreamTrack: PluginMediaStreamTrack) -> Void)?
 	) {
 		NSLog("PluginMediaStream#setListener()")
 
@@ -104,7 +110,7 @@ class PluginMediaStream : NSObject {
 			return false
 		}
 
-		//onAddTrack(pluginMediaStreamTrack)
+		onAddTrack(pluginMediaStreamTrack)
 		return true
 	}
 
@@ -123,7 +129,7 @@ class PluginMediaStream : NSObject {
 			return false
 		}
 
-		//onRemoveTrack(pluginMediaStreamTrack)
+		onRemoveTrack(pluginMediaStreamTrack)
 		return true
 	}
 
@@ -155,7 +161,7 @@ class PluginMediaStream : NSObject {
 		}
 
 		if self.eventListener != nil {
-			self.eventListenerForRemoveTrack!(track.id)
+			self.eventListenerForRemoveTrack!(track)
 
 			self.eventListener!([
 				"type": "removetrack",