+++ src/PluginMediaStreamTrack.swift
@@ -3,7 +3,6 @@ import Foundation
 
 class PluginMediaStreamTrack : NSObject {
 	var rtcMediaStreamTrack: RTCMediaStreamTrack
-	var streamId: String
 	var id: String
 	var kind: String
 	var eventListener: ((_ data: NSDictionary) -> Void)?
@@ -11,26 +10,26 @@ class PluginMediaStreamTrack : NSObject {
 	var lostStates = Array<String>()
 	var renders: [String : PluginMediaStreamRenderer]
 
-	init(rtcMediaStreamTrack: RTCMediaStreamTrack, streamId: String) {
+	init(rtcMediaStreamTrack: RTCMediaStreamTrack, trackId: String? = nil) {
 		NSLog("PluginMediaStreamTrack#init()")
 
 		self.rtcMediaStreamTrack = rtcMediaStreamTrack
 
-		// Handle possible duplicate remote trackId with janus name
-		// See: https://github.com/cordova-rtc/cordova-plugin-iosrtc/issues/432
-		if (rtcMediaStreamTrack.trackId.starts(with: "janus")) {
+		if (trackId == nil) {
+			// Handle possible duplicate remote trackId with  janus or short duplicate name
+			// See: https://github.com/cordova-rtc/cordova-plugin-iosrtc/issues/432
 			self.id = rtcMediaStreamTrack.trackId + "_" + UUID().uuidString;
 		} else {
-			self.id = rtcMediaStreamTrack.trackId;
+			self.id = trackId!;
 		}
-
+		
 		self.kind = rtcMediaStreamTrack.kind
 		self.renders = [:]
-		self.streamId = streamId;
 	}
 
 	deinit {
 		NSLog("PluginMediaStreamTrack#deinit()")
+		stop()
 	}
 
 	func run() {
@@ -120,6 +119,12 @@ class PluginMediaStreamTrack : NSObject {
 
 		// Let's try setEnabled(false), but it also fails.
 		self.rtcMediaStreamTrack.isEnabled = false
+		
+		self.eventListener!([
+			"type": "statechange",
+			"readyState": "ended",
+			"enabled": self.rtcMediaStreamTrack.isEnabled ? true : false
+		])
 
 		for (_, render) in self.renders {
 			render.stop()