+++ src/iosrtcPlugin.swift
@@ -75,11 +75,13 @@ class iosrtcPlugin : CDVPlugin {
 	}
 
 	@objc(onReset) override func onReset() {
-		NSLog("iosrtcPlugin#onReset() | doing nothing")
+		NSLog("iosrtcPlugin#onReset() | cleanup")
+		cleanup();
 	}
 
 	@objc(onAppTerminate) override func onAppTerminate() {
-		NSLog("iosrtcPlugin#onAppTerminate() | doing nothing")
+		NSLog("iosrtcPlugin#onAppTerminate() | cleanup")
+		cleanup();
 	}
 
 	@objc(new_RTCPeerConnection:) func new_RTCPeerConnection(_ command: CDVInvokedUrlCommand) {
@@ -112,7 +114,9 @@ class iosrtcPlugin : CDVPlugin {
 				self.emit(command.callbackId, result: result!)
 			},
 			eventListenerForAddStream: self.saveMediaStream,
-			eventListenerForRemoveStream: self.deleteMediaStream
+			eventListenerForRemoveStream: self.deleteMediaStream,
+			eventListenerForAddTrack: self.saveMediaStreamTrack,
+			eventListenerForRemoveTrack: self.deleteMediaStreamTrack
 		)
 
 		// Store the pluginRTCPeerConnection into the dictionary.
@@ -377,18 +381,11 @@ class iosrtcPlugin : CDVPlugin {
 	@objc(RTCPeerConnection_removeTrack:) func RTCPeerConnection_removeTrack(_ command: CDVInvokedUrlCommand) {
 		let pcId = command.argument(at: 0) as! Int
 		let trackId = command.argument(at: 1) as! String
-		let streamId = command.argument(at: 2) as! String
 		let pluginRTCPeerConnection = self.pluginRTCPeerConnections[pcId]
-		let pluginMediaStream = self.pluginMediaStreams[streamId]
 		let pluginMediaStreamTrack = self.pluginMediaStreamTracks[trackId]
 
 		if pluginRTCPeerConnection == nil {
 			NSLog("iosrtcPlugin#RTCPeerConnection_removeTrack() | ERROR: pluginRTCPeerConnection with pcId=%@ does not exist", String(pcId))
-			return;
-		}
-
-		if pluginMediaStream == nil {
-			NSLog("iosrtcPlugin#RTCPeerConnection_removeTrack() | ERROR: pluginMediaStream with id=%@ does not exist", String(streamId))
 			return;
 		}
 
@@ -400,7 +397,7 @@ class iosrtcPlugin : CDVPlugin {
 		self.queue.async { [weak pluginRTCPeerConnection, weak pluginMediaStreamTrack] in
 			pluginRTCPeerConnection?.removeTrack(pluginMediaStreamTrack!)
 			// TODO remove only if not used by other stream
-			self.deleteMediaStreamTrack(trackId)
+			// self.deleteMediaStreamTrack(pluginMediaStreamTrack!)
 		}
 	}
 
@@ -690,7 +687,7 @@ class iosrtcPlugin : CDVPlugin {
 
 		if self.pluginMediaStreams[streamId] == nil {
 			let rtcMediaStream : RTCMediaStream = self.rtcPeerConnectionFactory.mediaStream(withStreamId: streamId)
-			let pluginMediaStream = PluginMediaStream(rtcMediaStream: rtcMediaStream)
+			let pluginMediaStream = PluginMediaStream(rtcMediaStream: rtcMediaStream, streamId: streamId)
 			pluginMediaStream.run()
 
 			self.saveMediaStream(pluginMediaStream)
@@ -794,6 +791,27 @@ class iosrtcPlugin : CDVPlugin {
 		self.pluginMediaStreams[id] = nil
 	}
 
+	@objc(MediaStreamTrack_clone:) func MediaStreamTrack_clone(_ command: CDVInvokedUrlCommand) {
+		NSLog("iosrtcPlugin#MediaStreamTrack_clone()")
+
+		let existingTrackId = command.argument(at: 0) as! String
+		let newTrackId = command.argument(at: 1) as! String
+		let pluginMediaStreamTrack = self.pluginMediaStreamTracks[existingTrackId]
+
+		if pluginMediaStreamTrack == nil {
+			NSLog("iosrtcPlugin#MediaStreamTrack_clone() | ERROR: pluginMediaStreamTrack with id=%@ does not exist", String(existingTrackId))
+			return;
+		}
+
+		if self.pluginMediaStreams[newTrackId] == nil {
+			let rtcMediaStreamTrack = self.pluginMediaStreamTracks[existingTrackId]!.rtcMediaStreamTrack;
+			let newPluginMediaStreamTrack = PluginMediaStreamTrack(rtcMediaStreamTrack: rtcMediaStreamTrack, trackId: newTrackId)
+
+			self.saveMediaStreamTrack(newPluginMediaStreamTrack)
+		} else {
+			NSLog("iosrtcPlugin#MediaStreamTrack_clone() | ERROR: pluginMediaStreamTrack with id=%@ already exist", String(newTrackId))
+		}
+	}
 
 	@objc(MediaStreamTrack_setListener:) func MediaStreamTrack_setListener(_ command: CDVInvokedUrlCommand) {
 		NSLog("iosrtcPlugin#MediaStreamTrack_setListener()")
@@ -821,7 +839,7 @@ class iosrtcPlugin : CDVPlugin {
 				},
 				eventListenerForEnded: { () -> Void in
 					// Remove the track from the container.
-					self.pluginMediaStreamTracks[pluginMediaStreamTrack!.id] = nil
+					self.deleteMediaStreamTrack(pluginMediaStreamTrack!);
 				}
 			)
 		}
@@ -1058,6 +1076,12 @@ class iosrtcPlugin : CDVPlugin {
 		}
 	}
 
+	@objc(initAudioDevices:) func initAudioDevices(_ command: CDVInvokedUrlCommand) {
+		NSLog("iosrtcPlugin#initAudioDevices()")
+
+		PluginRTCAudioController.initAudioDevices()
+	}
+
 	@objc(RTCTurnOnSpeaker:) func RTCTurnOnSpeaker(_ command: CDVInvokedUrlCommand) {
 		DispatchQueue.main.async {
 			let isTurnOn: Bool = CBool(command.arguments[0] as! Bool)
@@ -1112,25 +1136,27 @@ class iosrtcPlugin : CDVPlugin {
 		if self.pluginMediaStreams[pluginMediaStream.id] == nil {
 			self.pluginMediaStreams[pluginMediaStream.id] = pluginMediaStream
 		} else {
-			NSLog("- PluginMediaStreams already exist [id:%@]", String(pluginMediaStream.id))
+			NSLog("iosrtcPlugin#saveMediaStream() | WARN: PluginMediaStream already exist [id:%@]", String(pluginMediaStream.id))
 			return;
 		}
 
 		// Store its PluginMediaStreamTracks' into the dictionary.
-		for (id, track) in pluginMediaStream.audioTracks {
-			if self.pluginMediaStreamTracks[id] == nil {
-				self.pluginMediaStreamTracks[id] = track
-			}
+		for (_, pluginMediaStreamTrack) in pluginMediaStream.audioTracks {
+			saveMediaStreamTrack(pluginMediaStreamTrack);
 		}
-		for (id, track) in pluginMediaStream.videoTracks {
-			if self.pluginMediaStreamTracks[id] == nil {
-				self.pluginMediaStreamTracks[id] = track
-			}
+
+		for (_, pluginMediaStreamTrack) in pluginMediaStream.videoTracks {
+			saveMediaStreamTrack(pluginMediaStreamTrack);
 		}
 	}
 
-	fileprivate func deleteMediaStream(_ id: String) {
-		self.pluginMediaStreams[id] = nil
+	fileprivate func deleteMediaStream(_ pluginMediaStream: PluginMediaStream) {
+		if (self.pluginMediaStreams[pluginMediaStream.id] != nil) {
+			self.pluginMediaStreams[pluginMediaStream.id] = nil
+			
+			// deinit should call stop by itself
+			//pluginMediaStream.stop();
+		}
 	}
 
 	fileprivate func saveMediaStreamTrack(_ pluginMediaStreamTrack: PluginMediaStreamTrack) {
@@ -1139,7 +1165,48 @@ class iosrtcPlugin : CDVPlugin {
 		}
 	}
 
-	fileprivate func deleteMediaStreamTrack(_ id: String) {
-		self.pluginMediaStreamTracks[id] = nil
+	fileprivate func deleteMediaStreamTrack(_ pluginMediaStreamTrack: PluginMediaStreamTrack) {
+		if (self.pluginMediaStreamTracks[pluginMediaStreamTrack.id] != nil) {
+			self.pluginMediaStreamTracks[pluginMediaStreamTrack.id] = nil
+			
+			// deinit should call stop by itself
+			//pluginMediaStreamTrack.stop();
+		}
+	}
+
+	fileprivate func cleanup() {
+
+		// Close all RTCPeerConnections
+		for (pcId, pluginRTCPeerConnection) in self.pluginRTCPeerConnections {
+			pluginRTCPeerConnection.close()
+			self.pluginRTCPeerConnections[pcId] = nil;
+		}
+
+		// Close all StreamRenderers
+		for (id, pluginMediaStreamRenderer) in self.pluginMediaStreamRenderers {
+			pluginMediaStreamRenderer.close()
+			self.pluginMediaStreamRenderers[id] = nil;
+		}
+
+		// Close All MediaStream
+		for (streamId, pluginMediaStream) in self.pluginMediaStreams {
+			// Store its PluginMediaStreamTracks' into the dictionary.
+			for (trackId, pluginMediaStreamTrack) in pluginMediaStream.audioTracks {
+				pluginMediaStream.removeTrack(pluginMediaStreamTrack);
+				deleteMediaStreamTrack(pluginMediaStreamTrack);
+			}
+
+			for (trackId, pluginMediaStreamTrack) in pluginMediaStream.videoTracks {
+				pluginMediaStream.removeTrack(pluginMediaStreamTrack);
+				deleteMediaStreamTrack(pluginMediaStreamTrack);
+			}
+
+			deleteMediaStream(pluginMediaStream);
+		}
+
+		// Close All MediaStreamTracks without MediaStream
+		for (trackId, pluginMediaStreamTrack) in self.pluginMediaStreamTracks {
+			deleteMediaStreamTrack(pluginMediaStreamTrack);
+		}
 	}
 }